% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{ACTG175}
\alias{ACTG175}
\title{ACTG175, block randmized study from speff2trial package}
\format{
Randomized study
}
\source{
Hammer et al. 1996, speff2trial package.
}
\description{
Data from speff2trial
}
\examples{
data(ACTG175)
}
\keyword{data}
