% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arts_overview.R
\name{arts_overview}
\alias{arts_overview}
\title{arts_overview
search artworks by defining parameters
All parameters require a quotation and case sensitive except "q" and numbers.}
\usage{
arts_overview(
  q,
  isOnView = NULL,
  medium = NULL,
  hasImages = NULL,
  geoLocation = NULL,
  dateBegin = NULL,
  dateEnd = NULL
)
}
\arguments{
\item{q}{The main query.}

\item{isOnView}{"true" or "false".}

\item{medium}{eg. "Ceramics", "Furniture", "Paintings"}

\item{hasImages}{"true" or "false"}

\item{geoLocation}{eg. "Europe", "France", "Paris", "China"}

\item{dateBegin}{Positive for A.D, negative for B.C}

\item{dateEnd}{Same as "dateBegin"}
}
\value{
A dataframe with 15 columns
The output is a tidy dataframe including columns "objectID", "objectName", "culture", ect.
}
\description{
arts_overview
search artworks by defining parameters
All parameters require a quotation and case sensitive except "q" and numbers.
}
