% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifreg.R
\name{cifregFG}
\alias{cifregFG}
\title{Fine-Gray Cumulative Incidence Function Regression}
\usage{
cifregFG(formula, data, propodds = NULL, ...)
}
\arguments{
\item{formula}{Formula with an 'Event' outcome.}

\item{data}{Data frame.}

\item{propodds}{Set to \code{NULL} to fit the Fine-Gray model (default behavior of this function).}

\item{...}{Additional arguments passed to \code{cifreg}.}
}
\value{
An object of class \code{"cifreg"} (extending \code{"phreg"}) with \code{propodds=NULL}.
}
\description{
Convenience wrapper for \code{cifreg} that specifically fits the Fine-Gray model
by setting \code{propodds=NULL}. This provides subdistribution hazard ratio interpretations.
}
\seealso{
\code{\link{cifreg}}
}
\author{
Thomas Scheike
}
