% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterindex-reshape.R
\name{coarse_clust}
\alias{coarse_clust}
\title{Coarsen Cluster Identifiers}
\usage{
coarse_clust(clusters, max.clust = 100)
}
\arguments{
\item{clusters}{vector of cluster identifiers.}

\item{max.clust}{maximum number of coarsened clusters.}
}
\value{
Integer vector of coarsened cluster indices (0-based).
}
\description{
Reduces the number of unique clusters by binning into quantile groups.
}
