% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{CPH_HPN_CRBSI}
\alias{CPH_HPN_CRBSI}
\title{Rates for HPN program for patients of Copenhagen Cohort}
\format{
crbsi: cumulative rate of catheter related bloodstream infection in HPN patients of Copenhagen
 mechanical: cumulative rate of Mechanical (hole/defect) complication for catheter of HPN patients of Copenhagen
trombo: cumulative rate of Occlusion/Thrombosis complication for catheter of HPN patients of Copenhagen
terminal: rate of terminal event, patients leaving the HPN program
}
\source{
Estimated data
}
\description{
Rates for HPN program for patients of Copenhagen Cohort
}
\keyword{data}
