% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.conquer.R
\name{divide_conquer}
\alias{divide_conquer}
\title{Split a data set and run function}
\usage{
divide_conquer(func = NULL, data, size, splits, id = NULL, ...)
}
\arguments{
\item{func}{called function}

\item{data}{data-frame}

\item{size}{size of splits}

\item{splits}{number of splits (ignored if size is given)}

\item{id}{optional cluster variable}

\item{...}{Additional arguments to lower level functions}
}
\description{
Split a data set and run function
}
\examples{
## avoid dependency on timereg
## library(timereg)
## data(TRACE)
## res <- divide_conquer(prop.odds,TRACE,
## 	     formula=Event(time,status==9)~chf+vf+age,n.sim=0,size=200)
}
\author{
Thomas Scheike, Klaus K. Holst
}
