% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleFGR.r
\name{doubleFGR}
\alias{doubleFGR}
\alias{phreg}
\alias{phreg.par}
\alias{robust.phreg}
\alias{readPhreg}
\alias{bplotdFG}
\alias{predictdFG}
\title{Double CIF Fine-Gray model with two causes}
\usage{
doubleFGR(formula, data, offset = NULL, weights = NULL, X2 = NULL, ...)
}
\arguments{
\item{formula}{formula with 'Event'}

\item{data}{data frame}

\item{offset}{offsets for cox model}

\item{weights}{weights for Cox score equations}

\item{X2}{specifies the regression design for second CIF model}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Estimation based on derived hazards and recursive esitmating euqations.
fits two parametrizations 
1) 
 \deqn{
F_1(t,X) = 1 - \exp( \exp( X^T \beta ) \Lambda_1(t)) 
}
and 
 \deqn{
F_2(t,X_2) = 1 - \exp( \exp( X_2^T \beta_2 ) \Lambda_2(t)) 
}
or restricted version
2) 
 \deqn{
F_1(t,X) = 1 - \exp( \exp( X^T \beta ) \Lambda_1(t)) 
}
and 
 \deqn{
F_2(t,X_2,X) = ( 1 - \exp( \exp( X_2^T \beta_2 ) \Lambda_2(t)) ) (1 - F_1(\infty,X))
}
}
\examples{
res <- 0 
data(bmt)
bmt$age2 <- bmt$age 
newdata <- bmt[1:19,]
par(mfrow=c(5,3))

## same X1 and X2
pr2 <- doubleFGR(Event(time,cause)~age+platelet,data=bmt,restrict=res)
bplotdFG(pr2,cause=1)
bplotdFG(pr2,cause=2,add=TRUE)
pp21 <- predictdFG(pr2,newdata=newdata)
pp22 <- predictdFG(pr2,newdata=newdata,cause=2)
plot(pp21)
plot(pp22,add=TRUE,col=2)
pp21 <- predictdFG(pr2)
pp22 <- predictdFG(pr2,cause=2)
plot(pp21)
plot(pp22,add=TRUE,col=2)

pr2 <- doubleFGR(Event(time,cause)~strata(platelet),data=bmt,restrict=res)
bplotdFG(pr2,cause=1)
bplotdFG(pr2,cause=2,add=TRUE)
pp21 <- predictdFG(pr2,newdata=newdata)
pp22 <- predictdFG(pr2,,newdata=newdata,cause=2)
plot(pp21)
plot(pp22,add=TRUE,col=2)
pp21 <- predictdFG(pr2)
pp22 <- predictdFG(pr2,cause=2)
plot(pp21)
plot(pp22,add=TRUE,col=2)

## different X1 and X2
pr2 <- doubleFGR(Event(time,cause)~age+platelet+age2,data=bmt,X2=3,restrict=res)
bplotdFG(pr2,cause=1)
bplotdFG(pr2,cause=2,add=TRUE)
pp21 <- predictdFG(pr2,newdata=newdata)
pp22 <- predictdFG(pr2,newdata=newdata,cause=2)
plot(pp21)
plot(pp22,add=TRUE,col=2)
pp21 <- predictdFG(pr2)
pp22 <- predictdFG(pr2,cause=2)
plot(pp21)
plot(pp22,add=TRUE,col=2)

### uden X1
pr2 <- doubleFGR(Event(time,cause)~age+platelet,data=bmt,X2=1:2,restrict=res)
bplotdFG(pr2,cause=1)
bplotdFG(pr2,cause=2,add=TRUE)
pp21 <- predictdFG(pr2,newdata=newdata)
pp22 <- predictdFG(pr2,newdata=newdata,cause=2)
plot(pp21)
plot(pp22,add=TRUE,col=2)
pp21 <- predictdFG(pr2)
p22 <- predictdFG(pr2,cause=2)
plot(pp21)
plot(pp22,add=TRUE,col=2)

### uden X2
pr2 <- doubleFGR(Event(time,cause)~age+platelet,data=bmt,X2=0,restrict=res)
bplotdFG(pr2,cause=1)
bplotdFG(pr2,cause=2,add=TRUE)
pp21 <- predictdFG(pr2,newdata=newdata)
pp22 <- predictdFG(pr2,newdata=newdata,cause=2)
plot(pp21)
plot(pp22,add=TRUE,col=2)
pp21 <- predictdFG(pr2)
pp22 <- predictdFG(pr2,cause=2)
plot(pp21)
plot(pp22,add=TRUE,col=2)

}
\author{
Thomas Scheike
}
