% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_binreg.R
\name{estimate.binregCV}
\alias{estimate.binregCV}
\title{Joint estimate across CV models for a binregCV object}
\usage{
\method{estimate}{binregCV}(x, f = NULL, ...)
}
\arguments{
\item{x}{A `binregCV` object.}

\item{f}{Optional transformation function passed to `lava::estimate`.}

\item{...}{Not currently used.}
}
\description{
Joint estimate across CV models for a binregCV object
}
