% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreg.R
\name{evalTerminal}
\alias{evalTerminal}
\title{Evaluates piece constant covariates at min(D,t) where D is a terminal event}
\usage{
evalTerminal(formula, data = data, death.code = 2, time = NULL)
}
\arguments{
\item{formula}{formula with 'Event' outcome}

\item{data}{data frame}

\item{death.code}{codes for death (terminating event, 2 default)}

\item{time}{for evaluation}
}
\description{
returns X(min(D,t)) and min(D,t) and their ratio. for censored observation 0. 
to use with the IPCW models implemented.
}
\author{
Thomas Scheike
}
