% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterindex-reshape.R
\name{faster.reshape}
\alias{faster.reshape}
\title{Fast Reshape from Long to Wide Format}
\usage{
faster.reshape(data, clusters, index.type = FALSE, num = NULL, Rindex = 1)
}
\arguments{
\item{data}{a matrix or data.frame to reshape.}

\item{clusters}{vector of cluster identifiers, or column name in data.}

\item{index.type}{logical; if TRUE, clusters are already 0-based indices.}

\item{num}{optional within-cluster numbering variable.}

\item{Rindex}{if 1, use R (1-based) indexing.}
}
\value{
A wide-format matrix.
}
\description{
Reshapes clustered long-format data to wide format efficiently using
compiled code.
}
