% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.same.cens.R
\name{force.same.cens}
\alias{force.same.cens}
\alias{force_same_cens}
\title{Force Same Censoring Within Clusters}
\usage{
force.same.cens(
  data,
  id = "id",
  time = "time",
  cause = "cause",
  entrytime = NULL,
  cens.code = 0
)

force_same_cens(
  data,
  id = "id",
  time = "time",
  cause = "cause",
  entrytime = NULL,
  cens.code = 0
)
}
\arguments{
\item{data}{a data.frame in long format.}

\item{id}{name of the cluster/pair identifier column.}

\item{time}{name of the time variable.}

\item{cause}{name of the cause/status variable.}

\item{entrytime}{optional name of left-truncation time variable.}

\item{cens.code}{value indicating censoring in the cause variable.}
}
\value{
A data.frame with enforced same censoring.
}
\description{
Enforces the same censoring time within clusters (pairs) by censoring
both subjects at the minimum censoring time when censoring occurs first.
}
