% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{ghaplos}
\alias{ghaplos}
\title{ghaplos  haplo-types for subjects of haploX data}
\source{
Simulated data
}
\description{
ghaplos  haplo-types for subjects of haploX data
}
\keyword{data}
