% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{haplo}
\alias{haplo}
\title{haplo fun data}
\format{
hapfreqs : haplo frequencies
haploX:  covariates and response for haplo survival discrete survival
ghaplos:  haplo-types for subjects of haploX data
}
\source{
Estimated data
}
\description{
haplo fun data
}
\keyword{data}
