% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{haploX}
\alias{haploX}
\title{haploX  covariates and response for haplo survival discrete survival}
\source{
Simulated data
}
\description{
haploX  covariates and response for haplo survival discrete survival
}
\keyword{data}
