% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{hfaction_cpx12}
\alias{hfaction_cpx12}
\title{hfaction, subset of block randmized study HF-ACtion from WA package}
\format{
Randomized study
}
\source{
WA package, Connor et al. 2009
}
\description{
Data from HF-action trial slightly modified from WA package
}
\examples{
data(hfaction_cpx12)
}
\keyword{data}
