% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{logrank_lt}
\alias{logrank_lt}
\title{Lu-Tsiatis More Efficient Log-Rank for Randomized studies with baseline covariates}
\usage{
logrank_lt(
  formula,
  data,
  augmentR = NULL,
  augmentC = NULL,
  km = TRUE,
  cens.code = 0,
  level = 0.95,
  cens.model = NULL,
  typeII = NULL,
  ...
)
}
\arguments{
\item{formula}{formula with 'Surv' outcome (see \code{coxph}) and treatment (randomization 0/1)}

\item{data}{data frame}

\item{augmentR}{formula for the randomization augmentation  (~age+sex)}

\item{augmentC}{formula for the censoring augmentation  (~age+sex)}

\item{km}{use Kaplan-Meier for the censoring weights (stratified on treatment)}

\item{cens.code}{censoring code}

\item{level}{of confidence intervals}

\item{typeII}{if 1 then computes also alternative formulae that are based on the censoring martingale rather than the robust processes of  Lu-Tsiatis computations.}

\item{...}{Additional arguments to phreg function}

\item{cens.model=NULL, }{default is censoring model ~strata(treatment) but any model can be used to make censoring martingales}
}
\description{
Efficient implementation of the Lu-Tsiatis improvement using baseline covariates. Results
almost equivalent with the speffSurv function of the speff2trial function. A dynamic 
censoring augmentation regression is also computed to gain even more from the censoring augmentation.
}
\examples{
## Lu, Tsiatis simulation
data <- mets:::simLT(0.7,200)

out <- logrank_lt(Surv(time,status)~Z,data=data,augmentR=~X,augmentC=~factor(Z):X)
out$coefs
}
\references{
Lu, Tsiatis (2008), Improving the efficiency of the log-rank test using auxiliary covariates, Biometrika, 679--694
}
\author{
Thomas Scheike
}
