% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediator-survival.R
\name{medweight}
\alias{medweight}
\title{Computes mediation weights}
\usage{
medweight(
  fit,
  data = data,
  var = NULL,
  name.weight = "weights",
  id.name = "id",
  ...
)
}
\arguments{
\item{fit}{either glm-binomial or mlogit (mets package)}

\item{data}{data frame with data}

\item{var}{is NULL reads mediator and exposure from formulae in the fit.}

\item{name.weight}{name of weights}

\item{id.name}{name of id variable, important for SE computations}

\item{...}{Additional arguments to}
}
\description{
Computes mediation weights for either binary or multinomial mediators.
The important part is that the influence functions can be obtained to compute standard errors.
}
\author{
Thomas Scheike
}
