% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{phregR}
\alias{phregR}
\alias{FastCoxPLstrataR}
\title{Fast Cox PH regression and calculations done in R to make play and adjustments easy}
\usage{
phregR(formula, data, offset = NULL, weights = NULL, ...)
}
\arguments{
\item{formula}{formula with 'Surv' outcome (see \code{coxph})}

\item{data}{data frame}

\item{offset}{offsets for cox model}

\item{weights}{weights for Cox score equations}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Fast Cox PH regression with R implementation to play and adjust in R function: FastCoxPLstrataR
}
\details{
Robust variance is default variance with the summary.

influence functions (iid) will follow numerical order of given cluster variable
so ordering after $id will give iid in order of data-set.
}
\author{
Klaus K. Holst, Thomas Scheike
}
