% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{plot.phreg}
\alias{plot.phreg}
\alias{basehazplotO.phreg}
\alias{baseplot}
\alias{bplot}
\alias{basecumhaz}
\alias{plotConfRegion}
\alias{plotConfRegionSE}
\alias{plotstrata}
\alias{kmplot}
\alias{plotConfregion}
\title{Plotting the baselines of stratified Cox}
\usage{
\method{plot}{phreg}(x, ...)
}
\arguments{
\item{x}{phreg object}

\item{...}{Additional arguments to baseplot funtion}
}
\description{
Plotting the baselines of stratified Cox
}
\examples{
data(TRACE)
dcut(TRACE) <- ~.
out1 <- phreg(Surv(time,status==9)~vf+chf+strata(wmicat.4),data=TRACE)

par(mfrow=c(2,2))
plot(out1)
plot(out1,stratas=c(0,3))
plot(out1,stratas=c(0,3),col=2:3,lty=1:2,se=TRUE)
plot(out1,stratas=c(0),col=2,lty=2,se=TRUE,polygon=FALSE)
plot(out1,stratas=c(0),col=matrix(c(2,1,3),1,3),lty=matrix(c(1,2,3),1,3),se=TRUE,polygon=FALSE)
}
\author{
Klaus K. Holst, Thomas Scheike
}
