% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_twin.R
\name{plot_twin}
\alias{plot_twin}
\title{Scatter plot function}
\usage{
plot_twin(
  data,
  density.args = list(),
  kernsmooth.args = list(),
  xlab,
  ylab,
  col = "black",
  col2 = "lightblue",
  alpha = 0.3,
  grid = TRUE
)
}
\arguments{
\item{data}{bivariate data to plot (data.frame or matrix with 2 columns)}

\item{density.args}{argumnets to (1d) density estimator}

\item{kernsmooth.args}{arguments to 2d-kernel smoother}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{col}{color of points}

\item{col2}{color of contour / density plot}

\item{alpha}{transparency level of points}

\item{grid}{should grid be added to the plot}
}
\description{
Scatterplot with contours of the (kernel) estimated density
}
\author{
Klaus Kähler Holst
}
