% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutils.R
\name{predictGLM}
\alias{predictGLM}
\title{Simple prediction based on glm with binomial link}
\usage{
predictGLM(object, newdata, id = NULL, fun = NULL, link.conf = TRUE, ...)
}
\arguments{
\item{object}{glm output}

\item{newdata}{possible newdata}

\item{id}{possible id for cluster corrected standard errors}

\item{fun}{possible function for non-standard predictions based on object}

\item{link.conf}{link transformation used}

\item{...}{arguments of estimate of lava for example level=0.95}
}
\description{
Simple prediction based on glm with binomial link
}
\author{
Thomas Scheike
}
