% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskRegression.R
\name{predictRisk.phreg}
\alias{predictRisk.phreg}
\alias{predictRisk.cifreg}
\alias{predictRisk.binreg}
\alias{predictRisk.recreg}
\title{Risk predictions to work with riskRegression package}
\usage{
\method{predictRisk}{phreg}(object, newdata, times = NULL, ...)
}
\arguments{
\item{object}{phreg/binreg/cifreg object}

\item{newdata}{newdata}

\item{times}{times for predictions}

\item{...}{additional arguments to lower level functions}
}
\description{
Risk predictions to work with riskRegression package
}
\author{
Thomas Scheike
}
