% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskRegression.R
\name{predictRisk}
\alias{predictRisk}
\title{Risk predictions to work with riskRegression package}
\usage{
predictRisk(object, ...)
}
\arguments{
\item{object}{phreg/binreg/cifreg object}

\item{...}{additional arguments to lower level functions}
}
\description{
Risk predictions to work with riskRegression package
}
\author{
Thomas Scheike
}
