% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_binreg.R
\name{print.summary_binregCV_multi}
\alias{print.summary_binregCV_multi}
\title{Print method for summary_binregCV_multi objects}
\usage{
\method{print}{summary_binregCV_multi}(x, digits = 4, ...)
}
\arguments{
\item{x}{A `summary_binregCV_multi` object.}

\item{digits}{Number of digits to round to.}

\item{...}{Not currently used.}
}
\description{
Print method for summary_binregCV_multi objects
}
