% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pch.R
\name{rchazC}
\alias{rchazC}
\title{Piecewise constant hazard distribution}
\usage{
rchazC(base1, rr, entry)
}
\arguments{
\item{base1}{baseline}

\item{rr}{relative risk terms}

\item{entry}{entry times for left truncation}
}
\description{
Piecewise constant hazard distribution
}
