% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{iid}
\alias{IC}
\alias{twostage}
\alias{estimate}
\alias{gof}
\alias{strata}
\alias{Surv}
\alias{cluster}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lava}{\code{\link[lava:estimate]{estimate()}}, \code{\link[lava:gof]{gof()}}, \code{\link[lava:IC]{IC()}}, \code{\link[lava:iid]{iid()}}, \code{\link[lava:twostage]{twostage()}}}

  \item{survival}{\code{\link[survival:cluster]{cluster()}}, \code{\link[survival:strata]{strata()}}, \code{\link[survival:Surv]{Surv()}}}
}}

