% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{resmean.phreg}
\alias{resmean.phreg}
\title{Restricted mean for stratified Kaplan-Meier with martingale standard errors}
\usage{
resmean.phreg(x, times = NULL, covs = NULL, ...)
}
\arguments{
\item{x}{phreg object}

\item{times}{possible times for which to report restricted mean}

\item{covs}{possible covariaates for Cox model}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Restricted mean for stratified Kaplan-Meier with martingale standard errors
}
\examples{
data(TRACE)
out1 <- phreg(Surv(time,status==9)~strata(vf,chf),data=TRACE)

rm1 <- resmean.phreg(out1,times=5)
summary(rm1)
mets:::plot.resmean_phreg(rm1)
}
\author{
Thomas Scheike
}
