% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pch.R
\name{rpch}
\alias{rpch}
\alias{ppch}
\title{Piecewise constant hazard distribution}
\usage{
rpch(n, lambda = 1, breaks = c(0, Inf))
}
\arguments{
\item{n}{sample size}

\item{lambda}{rate parameters}

\item{breaks}{time cut-points}
}
\description{
Piecewise constant hazard distribution
}
\examples{

}
