% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-pc-hazard.R
\name{sim.phregs}
\alias{sim.phregs}
\title{Simulation of cause specific from Cox models.}
\usage{
sim.phregs(
  coxs,
  n,
  data = NULL,
  rr = NULL,
  strata = NULL,
  entry = NULL,
  extend = NULL,
  cens = NULL,
  rrc = NULL,
  ...
)
}
\arguments{
\item{coxs}{list of cox models.}

\item{n}{number of simulations.}

\item{data}{to extract covariates for simulations (draws from observed covariates).}

\item{rr}{possible vector of relative risk for cox model.}

\item{strata}{possible vector of strata}

\item{entry}{delayed entry variable for simulation.}

\item{extend}{to extend possible stratified baselines to largest end-point}

\item{cens}{specifies censoring model, if NULL then only censoring for 
  each cause at end of last event of this type. 
if "is.matrix" then uses cumulative. 
 hazard given, if "is.scalar" then uses rate for exponential, and if not
 given then takes average rate of in simulated data from cox model.
 But censoring can also be given as a cause.}

\item{rrc}{possible vector of relative risk for cox-type censoring.}

\item{...}{arguments for rchaz, for example entry-time}
}
\description{
Simulates data that looks like fit from cause specific Cox models. 
Censor data automatically. When censoring is given in the  list of causes this
will give censoring that looks like the data.  Covariates are drawn from data-set
with replacement. This gives covariates like the data.  Calls sim.phregs
}
\examples{
library(mets)
data(bmt)
nsim <- 100; 

cox1 <- phreg(Surv(time,cause==1)~strata(tcell)+platelet+age,data=bmt)
cox2 <- phreg(Surv(time,cause==2)~tcell+strata(platelet),data=bmt)
coxs <- list(cox1,cox2)
## just calls sim.phregs !
dd <- sim.phregs(coxs,nsim,data=bmt,extend=c(0.001))
scox1 <- phreg(Surv(time,status==1)~strata(tcell)+platelet+age,data=dd)
scox2 <- phreg(Surv(time,status==2)~tcell+strata(platelet),data=dd)

cbind(cox1$coef,scox1$coef)
cbind(cox2$coef,scox2$coef)
par(mfrow=c(1,2))
plot(cox1); plot(scox1,add=TRUE); 
plot(cox2); plot(scox2,add=TRUE); 

}
\author{
Thomas Scheike
}
\keyword{survival}
