% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutils.R
\name{summaryGLM}
\alias{summaryGLM}
\title{Reporting OR from glm with binomial link}
\usage{
summaryGLM(object, id = NULL, fun = NULL, ...)
}
\arguments{
\item{object}{glm output}

\item{id}{possible id for cluster corrected standard errors}

\item{fun}{possible function for non-standard predictions based on object}

\item{...}{arguments of estimate of lava for example level=0.95}
}
\description{
Reporting OR from glm with binomial link
}
\author{
Thomas Scheike
}
