% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.boxarea.R
\name{surv_boxarea}
\alias{surv_boxarea}
\title{Bivariate Survival Data on Rectangular Regions}
\usage{
surv_boxarea(
  left.trunc,
  right.cens,
  data,
  timevar = "time",
  status = "status",
  id = "id",
  covars = NULL,
  covars.pairs = NULL,
  num = NULL,
  silent = 1,
  boxtimevar = "boxtime"
)
}
\arguments{
\item{left.trunc}{vector of length 2 giving left truncation times.}

\item{right.cens}{vector of length 2 giving right censoring times.}

\item{data}{a data.frame with the survival data.}

\item{timevar}{name of the time variable.}

\item{status}{name of the status variable.}

\item{id}{name of the cluster identifier.}

\item{covars}{optional covariate names.}

\item{covars.pairs}{optional pair-level covariate names.}

\item{num}{within-cluster numbering variable name.}

\item{silent}{verbosity level (1=silent).}

\item{boxtimevar}{name for the created box-time variable.}
}
\value{
A data.frame in long format restricted to the specified region.
}
\description{
Restricts bivariate survival data to a rectangular time region defined
by left-truncation and right-censoring boundaries, for use with
piecewise twostage models.
}
