% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA-estimands.R
\name{WA_reg}
\alias{WA_reg}
\title{While-Alive regression for recurrent events}
\usage{
WA_reg(
  formula,
  data,
  time = NULL,
  cens.code = 0,
  cause = 1,
  death.code = 2,
  marks = NULL,
  ...,
  trans = 1
)
}
\arguments{
\item{formula}{Event formula with regression design}

\item{data}{data frame}

\item{time}{for estimation}

\item{cens.code}{of censorings}

\item{cause}{of events}

\item{death.code}{of terminal events}

\item{marks}{possible marks for for recurrent events}

\item{...}{arguments for binreg}

\item{trans}{possible power for mean of events per time-unit, default is 1}
}
\description{
Does regression for the mean of the events per time unit \deqn{ Z(t)= N(min(D,t))/min(D,t))} based on
IPCW etimation calling the binreg after constructing the outcome \deqn{Z(t)}
}
\examples{
data(hfactioncpx12)
hfactioncpx12$age <- rnorm(741)[hfactioncpx12$id] 

dtable(hfactioncpx12,~status)
## exp-link regression 
dd <- WA_reg(Event(entry,time,status)~treatment+age+cluster(id),data=hfactioncpx12,
                    time=2,death.code=2)
summary(dd)
}
\references{
Nonparametric estimation of the Patient Weighted While-Alive Estimand arXiv preprint by A. Ragni, T. Martinussen, T. Scheike
}
\author{
Thomas Scheike
}
