% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutils.R
\name{waldTest}
\alias{waldTest}
\title{Wald test for model (type III test)}
\usage{
waldTest(object, ...)
}
\arguments{
\item{object, }{for example glm object that can be used with estimate}

\item{...}{arguments for estimate of lava for example id=data$id for cluster correction}
}
\description{
Wald test for model (type III test)
}
\author{
Thomas Scheike
}
