% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arts_images.R
\name{arts_images}
\alias{arts_images}
\title{arts_images
extract the primary picture of artworks matching parameters
Input the objectIDs from the previous function that you're interested}
\usage{
arts_images(objectIDs)
}
\arguments{
\item{objectIDs}{A 6 digit number. Using c(ID1, ID2) for multiple IDs}
}
\value{
A table with the object ID, name, and photo
}
\description{
arts_images
extract the primary picture of artworks matching parameters
Input the objectIDs from the previous function that you're interested
}
