% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arts_overview.R
\name{arts_overview}
\alias{arts_overview}
\title{arts_overview
compare the change 1st.
compare the change 1st.
This is a function mainly used to find masterpieces you may be interested by defining several parameters.
The parameters consists of "q", "isOnView", "medium", "hasImages", "geoLocation", "dateBegin", and "dateEnd".
The output is a tidy dataframe including columns "objectID", "objectName", "culture", "period", "reign", "artistDisplayName", "artistDisplayBio", "medium", "dimensions", "city", "country", "region", "excavation", "classification", and "objectURL".}
\usage{
arts_overview(
  q,
  isOnView = NULL,
  medium = NULL,
  hasImages = NULL,
  geoLocation = NULL,
  dateBegin = NULL,
  dateEnd = NULL
)
}
\description{
arts_overview
compare the change 1st.
compare the change 1st.
This is a function mainly used to find masterpieces you may be interested by defining several parameters.
The parameters consists of "q", "isOnView", "medium", "hasImages", "geoLocation", "dateBegin", and "dateEnd".
The output is a tidy dataframe including columns "objectID", "objectName", "culture", "period", "reign", "artistDisplayName", "artistDisplayBio", "medium", "dimensions", "city", "country", "region", "excavation", "classification", and "objectURL".
}
