% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{data}
\name{calgb8923}
\alias{calgb8923}
\title{CALGB 8923, twostage randomization SMART design}
\format{
Data from smart design
}
\source{
https://github.com/ycchao/code_Joint_model_SMART
}
\description{
Data from CALGB 8923
}
\examples{
data(calgb8923)
}
\keyword{data}
