% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{cif_yearslost}
\alias{cif_yearslost}
\title{Restricted mean time lost for competing risks with martingale standard errors}
\usage{
cif_yearslost(formula, data = data, cens.code = 0, times = NULL, ...)
}
\arguments{
\item{formula}{for phreg object with strata to inidicate strata or +1 if no strata is given}

\item{data}{frame for calculations}

\item{cens.code}{censoring code (needed to separate event codes from censorings)}

\item{times}{possible times for which to report restricted mean, summary will display the estimates}

\item{...}{Additional arguments to phreg}
}
\description{
Restricted mean time lost for competing risks based on integrated Aalen-Johansen.
A set of time points can be given to be returned in the summary,
but the function compuates years-lost for all event times and can be plotted/viewed.  
The RMTL for a specific time-point can be computed by using the rmstIPCW function and
the influence functions for the estimator is available.
}
\examples{
data(bmt); bmt$time <- bmt$time+runif(408)*0.001

## years.lost decomposed into causes
drm1 <- cif_yearslost(Event(time,cause)~strata(tcell,platelet),data=bmt,times=c(40,50))
par(mfrow=c(1,2)); plot(drm1,cause=1,se=1); plot(drm1,cause=2,se=1);
summary(drm1)
estimate(drm1,cause=1)
estimate(drm1,cause=2)

## comparing populations, can also be done using rmstIPCW via influence functions
drm1 <- cif_yearslost(Event(time,cause)~strata(tcell,platelet),data=bmt,times=40)
summary(drm1,contrast=list(1:4))
## first cause
e1 <- estimate(drm1)
estimate(e1,rbind(c(1,-1,0,0)))
}
\author{
Thomas Scheike
}
