% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-survival-haplo.r
\name{haplo.surv.discrete}
\alias{haplo.surv.discrete}
\title{Discrete time to event haplo type analysis}
\usage{
haplo.surv.discrete(X = NULL, y = "y", time.name = "time",
  Haplos = NULL, id = "id", desnames = NULL, designfunc = NULL,
  beta = NULL, no.opt = FALSE, method = "NR", stderr = TRUE,
  designMatrix = NULL, covnames = NULL, ...)
}
\arguments{
\item{X}{}

\item{y}{name of response (binary response with logistic link)}

\item{time.name}{(to construct design ~factor(time))}

\item{Haplos}{(data.frame with id, haplo1, haplo2 (haplotypes (h)) and  P(h|G))}

\item{id}{name of id variale}

\item{designfunc}{function that computes x(h)}

\item{beta}{}

\item{no.opt}{optimization TRUE/FALSE}

\item{method}{NR, nlm}

\item{designMatrix}{}

\item{covnames}{names of}

\item{...}{Additional arguments to lower level funtions lava:::NR  optimizer or nlm}

\item{des.names}{names for design matrix}
}
\description{
Can be used for logistic regression when time variable is "1" for all.
}
\details{
Cycle-specific logistic regression of haplo-type effects with known 
haplo-type probabilities. Given observed genotype G and unobserved haplotypes H
we here mix out over the possible haplotypes using that P(H|G) is provided.

\deqn{
S(t|x,G)) = E( S(t|x,H) | G)  = \sum_{h \in G} P(h|G) S(t|z,h) 
}
so survival can be computed by mixing out over possible h given g.

Survival is based on logistic regression for the discrete hazard function of the
form 
\deqn{
logit(P(T=t| T \geq t, x,h)) = \alpha_t + x(h) \beta
}
where x(h) is a regression design of x and haplotypes \eqn{h=(h_1,h_2)}

Take all names that begins with "haplo" and uses for design and sub-grouping (the h's)
}
\author{
Thomas Scheike
}
