% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{iidBaseline}
\alias{iidBaseline}
\title{Influence functions or IID decomposition of baseline for recrec/phreg/cifregFG}
\usage{
iidBaseline(
  object,
  time = NULL,
  ft = NULL,
  fixbeta = NULL,
  beta.iid = NULL,
  tminus = FALSE,
  ...
)
}
\arguments{
\item{object}{phreg/recreg/cifregFG object}

\item{time}{for baseline IID}

\item{ft}{function to compute IID of baseline integrated against f(t)}

\item{fixbeta}{to fix the coefficients}

\item{beta.iid}{to use these iid of beta}

\item{tminus}{to get predictions in t-}

\item{...}{additional arguments to lower level functions}
}
\description{
Influence functions or IID decomposition of baseline for recrec/phreg/cifregFG
}
\author{
Thomas Scheike
}
