% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\name{npc}
\alias{npc}
\alias{plotcr}
\alias{nonparcuminc}
\alias{sim_nordic}
\alias{corsim_prostate}
\alias{alpha2kendall}
\alias{alpha2spear}
\alias{coefmat}
\alias{piecewise_twostage}
\alias{surv_boxarea}
\alias{faster.reshape}
\alias{piecewise_data}
\alias{sim_BinPlack}
\alias{sim_BinFam}
\alias{sim_BinFam2}
\alias{sim_SurvFam}
\alias{corsim_prostate_random}
\alias{sim_nordic_random}
\alias{sim_Cox}
\alias{sim}
\alias{grouptable}
\alias{jumptimes}
\alias{folds}
\alias{casewise_bin}
\alias{force.same.cens}
\alias{or_cif}
\alias{Grandom.cif}
\alias{conftype}
\alias{kendall.ClaytonOakes.twin.ace}
\alias{random.cif}
\alias{recurrentMarginal}
\alias{summarybase.phreg}
\alias{survival.twostage}
\alias{twin.polygen.design}
\alias{rmst_phreg}
\alias{rr_cif}
\alias{‘sim_binClaytonOakes_pairs’}
\alias{sim_bptwin}
\alias{sim_Compete_twin_ace}
\alias{‘sim_nordictwin’}
\alias{simrchaz}
\alias{ace_family_design}
\alias{ascertained_pairs}
\alias{summaryTimeobject}
\alias{tie_breaker}
\alias{coarse_clust}
\alias{concordanceTwinACE}
\alias{concordanceTwostage}
\alias{fast.cluster}
\alias{force_same_cens}
\alias{ilap}
\alias{kendall_ClaytonOakes_twin_ace}
\alias{kendall_normal_twin_ace}
\alias{make_pairwise_design}
\alias{make_pairwise_design_competing}
\alias{matplot.mets.twostage}
\alias{object.defined}
\alias{p11.binomial.twostage.RV}
\alias{predictPairPlack}
\alias{sim_binClaytonOakes_family_ace}
\alias{sim_binClaytonOakes.pairs}
\alias{sim_binClaytonOakes_twin_ace}
\alias{sim_ClaytonOakes_family_ace}
\alias{sim_ClaytonOakes_twin_ace}
\alias{sim_Frailty_simple}
\alias{sim_binClaytonOakes_pairs}
\alias{sim_nordictwin}
\alias{sim_Compete_simple}
\alias{simCompete_twin_ace}
\alias{twin_polygen_design}
\alias{procform}
\alias{procform3}
\alias{procformdata}
\alias{drop.specials}
\alias{cumsumstrata}
\alias{matdoubleindex}
\alias{mdi}
\alias{revcumsum}
\alias{revcumsumstrata}
\alias{robust.basehaz.phreg}
\alias{sumstrata}
\alias{strataAugment}
\alias{predicttime}
\title{For internal use}
\description{
For internal use
}
\author{
Klaus K. Holst
}
\keyword{utilities}
