% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{km}
\alias{km}
\title{Kaplan-Meier with robust standard errors}
\usage{
km(formula, data = data, ...)
}
\arguments{
\item{formula}{formula with 'Surv' 'Event' outcome}

\item{data}{data frame}

\item{...}{Additional arguments to phreg}
}
\description{
Kaplan-Meier with robust standard errors 
Robust variance is default variance and obtained from the predict call
}
\examples{
library(mets)
data(TRACE)
TRACE$cluster <- sample(1:100,1878,replace=TRUE)
out1 <- km(Surv(time,status==9)~strata(vf,chf),data=TRACE)
out2 <- km(Surv(time,status==9)~strata(vf,chf)+cluster(cluster),data=TRACE)
dd <- expand.grid(vf=0:1,chf=0:1)

pout1 <- predict(out1,dd,robust=TRUE)
pout2 <- predict(out1,dd,robust=TRUE)

par(mfrow=c(1,2))
plot(pout1,se=TRUE)
plot(pout2,se=TRUE)
}
\author{
Thomas Scheike
}
