% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent.marginal.R
\name{logrankRecurrent}
\alias{logrankRecurrent}
\alias{logrankRecurrentBase}
\title{logrank type test for recurrent marginal mean or cumulative incidence}
\usage{
logrankRecurrent(
  recurrent,
  death,
  weight = c("I", "II"),
  km = TRUE,
  start = 0,
  stop = NULL,
  at.risk = 5,
  cluster.id = NULL
)
}
\arguments{
\item{recurrent}{phreg object}

\item{death}{phreg object}

\item{weight}{"I" or "II"}

\item{km}{to use Kaplan-Meier rather than exp(-cumhaz)}

\item{start}{integral start}

\item{stop}{integral stop, default max jumptime}

\item{at.risk}{only uses time-points where at most at.risk under risk}

\item{cluster.id}{additional cluster to combine iid representation that is ordered after id given (so safest to order data after id, that then can be GEE clustered additionally)}
}
\description{
The logrank type test-statistic 
\deqn{ \int_0^t  w(t) [ d \hat m_1(s) - d \hat m_2(s) ] } is computed and returned
with a robust variance estimate based on the influence functions. Works for the
recurrent events and cumulative incidence and also with delayed entry.
The weights are either default  "I" \deqn{ w(t)= R_1 R_2 / (R_1+R_2)} 
or  "II" \deqn{ w(t)= Y_1 Y_2 / (Y_1+Y_2) }  with 
\deqn{ Y_j(t)}  the number of subjects under risk in each group and 
\deqn{ R_j(t)=Y_j(t)/\hat S_j(t-)}  where the denominator is the Kaplan-Meier of death.
}
\examples{
library(mets)
data(hfactioncpx12)
hf <- hfactioncpx12

##xr <- phreg(Surv(entry,time,status==1)~strata(treatment)+cluster(id),data=hf)
##dr <- phreg(Surv(entry,time,status==2)~strata(treatment)+cluster(id),data=hf)
##logrankRecurrent(xr,dr,stop=5)

## same as 
outN <- recurrentMarginal(Event(entry,time,status)~strata(treatment)+cluster(id),
                         data=hf,cause=1,death.code=2)
logrankRecurrent(outN)
}
\references{
Ghosh and Lin (2002) Nonparametric Analysis of Recurrent events and death, Biometrics, 554--562.
}
\author{
Thomas Scheike
}
