% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{package}
\name{mets-package}
\alias{mets}
\alias{mets-package}
\title{Analysis of Multivariate Events}
\description{
Implementation of various statistical models for multivariate
event history data. Including multivariate cumulative incidence models,
and bivariate random effects probit models (Liability models)
}
\examples{

## To appear

}
\seealso{
Useful links:
\itemize{
  \item \url{https://kkholst.github.io/mets/}
  \item Report bugs at \url{https://github.com/kkholst/mets/issues}
}

}
\author{
Klaus K. Holst and Thomas Scheike
}
\keyword{package}
