% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutinomialreg.R
\name{mlogit}
\alias{mlogit}
\title{Multinomial regression based on phreg regression}
\usage{
mlogit(formula, data, ref = NULL, ...)
}
\arguments{
\item{formula}{formula with outcome (see \code{coxph})}

\item{data}{data frame}

\item{ref}{to set reference category (not working yet)}

\item{...}{Additional arguments to lower level funtions}
}
\description{
returns also influence functions (possibly robust)
}
\examples{

data(bmt)
mreg <- mlogit(cause~tcell+platelet,bmt)
summary(mreg,type="martingale")


}
\author{
Thomas Scheike
}
