% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{phreg_IPTW}
\alias{phreg_IPTW}
\title{IPTW Cox, Inverse Probaibilty of Treatment Weighted Cox regression}
\usage{
phreg_IPTW(
  formula,
  data,
  treat.model = NULL,
  treat.var = NULL,
  weights = NULL,
  estpr = 1,
  pi0 = 0.5,
  se.cluster = NULL,
  ...
)
}
\arguments{
\item{formula}{for phreg}

\item{data}{data frame for risk averaging}

\item{treat.model}{propensity score model (binary or multinomial)}

\item{treat.var}{a 1/0 variable that indicates when propensity score is computed over time}

\item{weights}{may be given, and then uses weights*w(A) as the weights}

\item{estpr}{to estimate propensity scores and get infuence function contribution to uncertainty}

\item{pi0}{fixed simple weights}

\item{se.cluster}{to compute GEE type standard errors when additional cluster structure is present.}

\item{...}{arguments for phreg call}
}
\description{
Fits Cox model with treatment weights \deqn{ w(A)= \sum_a I(A=a)/\pi(a|X)}, where
 \deqn{\pi(a|X)=P(A=a|X)}. Computes
standard errors via influence functions that are returned as the IID argument. 
Propensity scores are fitted using either logistic regression (glm) or the multinomial model (mlogit) when more
than two categories for treatment. The treatment needs to be a factor and is identified on the rhs
of the "treat.model".  Recurrent events can be considered with start,stop structure and then cluster(id) must be
specified. Robust standard errors are computed.
}
\details{
Also works with cluster argument. Time-dependent propensity score weights can also be computed when treat.var is 1
and then at time of 2nd treatment (A_1) uses weights w_0(A_0) * w_1(A_1) where A_0 is first treatment.
}
\examples{
library(mets)
data <- mets:::simLT(0.7,100,beta=0.3,betac=0,ce=1,betao=0.3)
dfactor(data) <- Z.f~Z
out <- phreg_IPTW(Surv(time,status)~Z.f,data=data,treat.model=Z.f~X)
summary(out)
estimate(out)
}
\author{
Thomas Scheike
}
