% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{phregIPTW}
\alias{phregIPTW}
\title{IPTW Cox, Inverse Probaibilty of treatment weighted Cox regression}
\usage{
phregIPTW(
  formula,
  data,
  treat.model = NULL,
  weights = NULL,
  estpr = 1,
  pi0 = 0.5,
  ...
)
}
\arguments{
\item{formula}{for phreg}

\item{data}{data frame for risk averaging}

\item{treat.model}{propensity score model (binary or multinomial)}

\item{weights}{may be given, and then uses weights*w(A) as the weights}

\item{estpr}{to estimate propensity scores and get infuence function contribution to uncertainty}

\item{pi0}{fixed simple weights}

\item{...}{arguments for phreg call}
}
\description{
Fits Cox model with treatment weights \deqn{ w(A)= \sum_a I(A=a)/P(A=a|X) }, computes
standard errors via influence functions that are returned as the IID argument. 
Propensity scores are fitted using either logistic regression or the multinomial model when more
than two categories for treatment. The treatment needs to be a factor and is identified on the rhs
of the "treat.model".
}
\details{
Also works with cluster argument.
}
\examples{
data(bmt)
dfactor(bmt) <- tcell~tcell
out <- phregIPTW(Surv(time,cause==1)~tcell+platelet+age,bmt,tcell~platelet+age)
summary(out)

}
\author{
Thomas Scheike
}
