% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskRegression.R
\name{predictRisk}
\alias{predictRisk}
\alias{predictRisk.phreg}
\alias{predictRisk.binreg}
\alias{predictRisk.cifreg}
\alias{predictRisk.cifregFG}
\alias{predictRisk.recreg}
\title{Risk predictions to work with riskRegression package}
\usage{
predictRisk(object, ...)
}
\arguments{
\item{object}{phreg/binreg/cifreg object}

\item{...}{additional arguments to lower level functions}

\item{newdata}{newdata}

\item{times}{times for predictions}
}
\description{
Risk predictions to work with riskRegression package
}
\author{
Thomas Scheike
}
