% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent.marginal.R
\name{recurrent_marginal}
\alias{recurrent_marginal}
\alias{recurrent_marginalAIPCW}
\alias{recurrentMarginalPhreg}
\alias{iidRecurrent}
\title{Fast recurrent marginal mean when death is possible}
\usage{
recurrent_marginal(formula, data, cause = 1, ..., death.code = 2, test = FALSE)
}
\arguments{
\item{formula}{with Event object}

\item{data}{data frame for computation}

\item{cause}{of interest (1 default)}

\item{...}{Additional arguments to lower level funtions}

\item{death.code}{codes for death (terminating event, 2 default)}

\item{test}{to compute logrank type test, only for two groups}
}
\description{
Fast Marginal means of recurrent events using the Ghosh and Lin (2000) 
robust standard errors. Fitting two models for death and recurent events, respectively, these are
combined to give the estimator 
\deqn{ \int_0^t  S(u-) dR(u)} the mean number of recurrent events (or the cumulative incidence), here
\deqn{ S(u) } is the probability of survival for the baseline group, and 
\deqn{ dR(u)} is the hazard rate of an event among survivors for the baseline.
}
\details{
Assumes no ties in the sense that jump times needs to be unique, this is particularly so for the stratified version.
}
\examples{
data(hfactioncpx12)
hf <- hfactioncpx12
hf$x <- as.numeric(hf$treatment) 

##  to fit non-parametric models with just a baseline 
xr <- phreg(Surv(entry,time,status==1)~cluster(id),data=hf)
dr <- phreg(Surv(entry,time,status==2)~cluster(id),data=hf)
par(mfrow=c(1,3))
plot(dr,se=TRUE)
title(main="death")
plot(xr,se=TRUE)
### robust standard errors 
rxr <-  robust_phreg(xr,fixbeta=1)
plot(rxr,se=TRUE,robust=TRUE,add=TRUE,col=4)

## marginal mean of expected number of recurrent events 
## out <- recurrentMarginalPhreg(xr,dr)
## summary(out,times=1:5) 

## marginal mean using formula  
outN <- recurrent_marginal(Event(entry,time,status)~cluster(id),hf,cause=1,death.code=2)
plot(outN,se=TRUE,col=2,add=TRUE)
summary(outN,times=1:5) 

########################################################################
###   with strata     ##################################################
########################################################################
out <- recurrent_marginal(Event(entry,time,status)~strata(treatment)+cluster(id),
                         data=hf,cause=1,death.code=2,test=TRUE)
plot(out,se=TRUE,ylab="marginal mean",col=1:2)

attr(out,"logrank")

summary(out,times=1:5) 
########################################################################
### iid decomposition based on influence functions from Ghosh Lin (2000)
########################################################################
head(iid(outN,time=3))

}
\references{
Cook, R. J. and Lawless, J. F. (1997) Marginal analysis of recurrent events and a terminating event. Statist. Med., 16, 911–924.
Ghosh and Lin (2002) Nonparametric Analysis of Recurrent events and death, Biometrics, 554--562.

Ghosh and Lin (2002) Nonparametric Analysis of Recurrent events and death, Biometrics, 554--562.
}
\author{
Thomas Scheike
}
