% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{IC}
\alias{iid}
\alias{twostage}
\alias{estimate}
\alias{gof}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lava}{\code{\link[lava:estimate.default]{estimate}}, \code{\link[lava]{gof}}, \code{\link[lava]{IC}}, \code{\link[lava]{iid}}, \code{\link[lava]{twostage}}}
}}

