% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreg.R
\name{resmean.phreg}
\alias{resmean.phreg}
\title{Restricted mean for stratified Kaplan-Meier or Cox model with martingale standard errors}
\usage{
resmean.phreg(x, times = NULL, covs = NULL, ...)
}
\arguments{
\item{x}{phreg object}

\item{times}{possible times for which to report restricted mean}

\item{covs}{possible covariate for Cox model}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Restricted mean for stratified Kaplan-Meier or stratified Cox with martingale 
standard error. Standard error is computed using linear interpolation between 
standard errors at jump-times. Plots gives restricted mean at all times.
}
\examples{
data(sTRACE)
out1 <- phreg(Surv(time,status==9)~strata(vf,chf),data=sTRACE)

rm1 <- resmean.phreg(out1,times=5)
summary(rm1)
par(mfrow=c(1,2))
plot(d1,se=1)
plot(d1,years.lost=TRUE,se=1)
}
\author{
Thomas Scheike
}
