% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted.mean.R
\name{resmeanIPCW}
\alias{resmeanIPCW}
\title{Restricted IPCW mean for censored survival data}
\usage{
resmeanIPCW(
  formula,
  data,
  cause = 1,
  time = NULL,
  beta = NULL,
  offset = NULL,
  weights = NULL,
  cens.weights = NULL,
  cens.model = ~+1,
  se = TRUE,
  kaplan.meier = TRUE,
  cens.code = 0,
  no.opt = FALSE,
  method = "nr",
  model = "exp",
  augmentation = NULL,
  h = NULL,
  MCaugment = NULL,
  ...
)
}
\arguments{
\item{formula}{formula with outcome (see \code{coxph})}

\item{data}{data frame}

\item{cause}{cause of interest}

\item{time}{time of interest}

\item{beta}{starting values}

\item{offset}{offsets for partial likelihood}

\item{weights}{for score equations}

\item{cens.weights}{censoring weights}

\item{cens.model}{only stratified cox model without covariates}

\item{se}{to compute se's  based on IPCW}

\item{kaplan.meier}{uses Kaplan-Meier for IPCW in contrast to exp(-Baseline)}

\item{cens.code}{gives censoring code}

\item{no.opt}{to not optimize}

\item{method}{for optimization}

\item{model}{exp or linear}

\item{augmentation}{to augment binomial regression}

\item{h}{h for estimating equation}

\item{MCaugment}{iid of h and censoring model}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Simple version of comp.risk function of timereg for just one time-point thus fitting the model 
\deqn{E(T \leq t | X ) = expit( X^T beta) }
}
\details{
Based on binomial regresion IPCW response estimating equation: 
\deqn{ X ( \Delta (T \leq t)/G_c(T_i-) - expit( X^T beta)) = 0 }
for IPCW adjusted responses. 

Based on binomial regresion IPCW response estimating equation: 
\deqn{ h(X) X ( \Delta (T \leq t)/G_c(T_i-) - expit( X^T beta)) = 0 }
for IPCW adjusted responses where $h$ is given as an argument together with iid of censoring with h. 

variance is based on  \deqn{ \sum w_i^2 } also with IPCW adjustment, and naive.var is variance 
under known censoring model. 

Censoring model may depend on strata.
}
\examples{

data(bmt); bmt$time <- bmt$time+runif(nrow(bmt))*0.001
# logistic regresion with IPCW binomial regression 
out <- resmeanIPCW(Event(time,cause!=0)~tcell+platelet+age,bmt,
                cause=1,time=50,cens.model=~strata(platelet),model="exp")
summary(out)

### same as Kaplan-Meier for full censoring model 
bmt$int <- with(bmt,interaction(tcell,platelet))
out <- resmeanIPCW(Event(time,cause!=0)~-1+int,bmt,cause=1,time=30,
                            cens.model=~strata(platelet,tcell),model="lin")
summary(out)
out1 <- phreg(Surv(time,cause!=0)~strata(tcell,platelet),data=bmt)
rm1 <- resmean.phreg(out1,times=30)
summary(rm1)
}
\author{
Thomas Scheike
}
