% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent.marginal.R
\name{simMultistate}
\alias{simMultistate}
\alias{simMultistateII}
\title{Simulation of illness-death model}
\usage{
simMultistate(n, cumhaz, cumhaz2, death.cumhaz, death.cumhaz2, rr = NULL,
  rr2 = NULL, rd = NULL, rd2 = NULL, gap.time = FALSE,
  max.recurrent = 100, dependence = 0, var.z = 0.22,
  cor.mat = NULL, cens = NULL, ...)
}
\arguments{
\item{n}{number of id's}

\item{cumhaz}{cumulative hazard of recurrent events}

\item{cumhaz2}{cumulative hazard of recurrent events  of type 2}

\item{death.cumhaz}{cumulative hazard of death from state 1}

\item{death.cumhaz2}{cumulative hazard of death from state 2}

\item{rr}{relative risk adjustment for cumhaz}

\item{rr2}{relative risk adjustment for cumhaz2}

\item{rd}{relative risk adjustment for death.cumhaz}

\item{rd2}{relative risk adjustment for death.cumhaz2}

\item{gap.time}{if true simulates gap-times with specified cumulative hazard}

\item{max.recurrent}{limits number recurrent events to 100}

\item{var.z}{variance of random effects}

\item{cor.mat}{correlation matrix for var.z variance of random effects}

\item{cens}{rate of censoring exponential distribution}

\item{...}{Additional arguments to lower level funtions}
}
\description{
Simulation of illness-death model
}
\details{
simMultistate with same death intensity from states 1 and 2 
simMultistateII with different death intensities from states 1 and 2

Must give cumulative hazards on some time-range
}
\examples{
########################################
## getting some rates to mimick 
########################################
data(base1cumhaz)
data(base4cumhaz)
data(drcumhaz)
dr <- drcumhaz
dr2 <- drcumhaz
dr2[,2] <- 1.5*drcumhaz[,2]
base1 <- base1cumhaz
base4 <- base4cumhaz

iddata <- simMultistate(1000,base1,base1,dr,dr2)
dlist(iddata,.~id|id<3,n=0)
 
c3 <- phreg(Surv(start,stop,status==3)~+strata(from),iddata)
c1 <- phreg(Surv(start,stop,status==1)~+1,subset(iddata,from==2))
c2 <- phreg(Surv(start,stop,status==2)~+1,subset(iddata,from==1))
###
par(mfrow=c(1,3))
bplot(c3,main="rates 1-> 3 , 2->3")
lines(dr,col=1,lwd=2)
lines(dr2,col=2,lwd=2)
###
bplot(c1,main="rate 1->2")
lines(base1,lwd=2)
###
bplot(c2,main="rate 2->1")
lines(base1,lwd=2)
 
}
\author{
Thomas Scheike
}
