% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent.marginal.R
\name{simRecurrent}
\alias{simRecurrent}
\title{Simulation of recurrent events data based on cumulative hazards for event and death process}
\usage{
simRecurrent(
  n,
  cumhaz,
  death.cumhaz = NULL,
  r1 = NULL,
  rd = NULL,
  rc = NULL,
  ...
)
}
\arguments{
\item{n}{number of id's}

\item{cumhaz}{cumulative hazard of recurrent events}

\item{death.cumhaz}{cumulative hazard of death}

\item{r1}{potential relative risk adjustment of rate}

\item{rd}{potential relative risk adjustment of rate}

\item{rc}{potential relative risk adjustment of rate}

\item{...}{Additional arguments to simRecurrentList}
}
\description{
Simulation of recurrent events data based on cumulative hazards for event and death process
}
