% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent.marginal.R
\name{simRecurrent}
\alias{simRecurrent}
\title{Simulation of recurrent events data based on cumulative hazards for event and death process}
\usage{
simRecurrent(
  n,
  cumhaz,
  death.cumhaz = NULL,
  r1 = NULL,
  rd = NULL,
  rc = NULL,
  ...
)
}
\arguments{
\item{...}{Additional arguments to simRecurrentList}
}
\description{
Simulation of recurrent events data based on cumulative hazards for event and death process
}
