% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-mean-tests.R
\name{test_marginalMean}
\alias{test_marginalMean}
\title{Pepe-Mori test, score-test type test, proportionality test}
\usage{
test_marginalMean(
  formula,
  data,
  cause = 1,
  cens.code = 0,
  ...,
  death.code = 2,
  death.code.prop = NULL,
  time = NULL,
  beta = NULL
)
}
\arguments{
\item{formula}{with Event object}

\item{data}{data frame for computation}

\item{cause}{of interest (1 default)}

\item{cens.code}{censoring code}

\item{...}{Additional arguments to lower level funtions}

\item{death.code}{codes for death (terminating event, 2 default)}

\item{death.code.prop}{code for other causes of death for Fine-Gray regression model}

\item{time}{upper limit in pepe-mori and AUC integrals  otherwise max event time}

\item{beta}{for score test}
}
\description{
for competing risks data and recurrent events the pepe-mori test uses the weights suggested in Ghosh and Lin (2000)
}
\examples{
library(mets)
data(bmt,package="mets")
bmt$time <- bmt$time+runif(nrow(bmt))*0.01
bmt$id <- 1:nrow(bmt)
dcut(bmt) <- age.f~age
str(bmt)      
     
fg=cifregFG(Event(time,cause)~tcell,data=bmt,cause=1)
 
## computing tests for difference  for CIF
pmt <- test_marginalMean(Event(time,cause)~strata(tcell)+cluster(id),data=bmt,cause=1,
			 death.code=1:2,death.code.prop=2,cens.code=0)
summary(pmt) 
 
pmt$pepe.mori
pmt$RatioAUC
pmt$prop.test
## score test equialent to Gray's test but variance estimated differently 
pmt$score.test$p.logrank.robust
 
### age-groups  
pmt <- test_marginalMean(Event(time,cause)~strata(age.f)+cluster(id),data=bmt,cause=1,
			 death.code=1:2,death.code.prop=2,cens.code=0)
summary(pmt) 
 
## having a look at the cumulative incidences 
cifs <- cif(Event(time,cause)~strata(age.f)+cluster(id),data=bmt,cause=1)
plot(cifs) 
 
## recurrent events   
data(hfactioncpx12)
hf <- hfactioncpx12
pmt <- test_marginalMean(Event(entry,time,status)~strata(treatment)+cluster(id),data=hf,
			 cause=1,death.code=2,cens.code=0)
summary(pmt) 
 
}
\author{
Thomas Scheike
}
