% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA-estimands.R
\name{WA_recurrent}
\alias{WA_recurrent}
\title{While-Alive estimands for recurrent events}
\usage{
WA_recurrent(
  formula,
  data,
  time = NULL,
  cens.code = 0,
  cause = 1,
  death.code = 2,
  trans = NULL,
  cens.formula = NULL,
  augmentR = NULL,
  augmentC = NULL,
  type = NULL,
  ...
)
}
\arguments{
\item{formula}{Event formula first covariate on rhs must be a factor giving the treatment}

\item{data}{data frame}

\item{time}{for estimation}

\item{cens.code}{of censorings}

\item{cause}{of events}

\item{death.code}{of terminal events}

\item{trans}{possible power for mean of events per time-unit}

\item{cens.formula}{censoring model, default is to use strata(treatment)}

\item{augmentR}{covariates for model of mean ratio}

\item{augmentC}{covariates for censoring augmentation}

\item{type}{augmentation for call of binreg, when augmentC is given default is "I" and otherwise "II"}

\item{...}{arguments for binregATE}
}
\description{
Considers the ratio of means \deqn{E(N(min(D,t)))/E(min(D,t))} and the
the mean of the events per time unit \deqn{E(N(min(D,t))/min(D,t))} both based on
IPCW etimation. RMST estimator equivalent to Kaplan-Meier based estimator.
}
\author{
Thomas Scheike
}
